x40
git-svn-id: https://daisy-model.googlecode.com/svn/trunk@1572 5f5bf4bb-3343-0410-8278-4fa8eb3319dc

+++ scope_sources.h
@@ -34,7 +34,7 @@ class ScopeSources : public Scope
   // Content.
 private:
   const std::vector<Source*> source;
-  std::vector<size_t> index;
+  std::vector<int> index;
 public:
   Time now;
